requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
distlib==0.3.1
docopt==0.6.2
docutils==0.17.1
docx2python==1.27.1
filelock==3.0.12
future==0.18.2
gTTS==2.2.2
idna==2.10
isort==5.7.0
Js2Py==0.71
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.2
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.20.2
opencv-python==4.5.2.52
packaging==20.9
Pillow==8.2.0
pipenv==2020.11.15
pipwin==0.5.1
playsound==1.2.2
PyAudio @ file:///E:/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyglet==1.4.10
Pygments==2.9.0
pyjsparser==2.7.1
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
pytesseract==0.3.7
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
pywin32==300
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
wrapt==1.12.1
